Use a version stamp for the migration indicator
